Updated to correctly utilize generics to create subclassable object that can implement comparable
